// src/openapi/generated/models/gridpoint12h-forecast-period.ts
export interface Gridpoint12hForecastPeriod {
    /**
     * Sequential period number.
     */
    'number'?: number;
    /**
     * A textual identifier for the period. This value will not be present for hourly forecasts. 
     */
    'name'?: string;
    /**
     * The starting time that this forecast period is valid for.
     */
    'startTime'?: string;
    /**
     * The ending time that this forecast period is valid for.
     */
    'endTime'?: string;
    /**
     * Indicates whether this period is daytime or nighttime.
     */
    'isDaytime'?: boolean;
    'temperature'?: GridpointHourlyForecastPeriodTemperature;
    /**
     * The unit of the temperature value (Fahrenheit or Celsius). This property is deprecated. Future versions will indicate the unit within the quantitative value object for the temperature property. To make use of the future standard format now, set the \"forecast_temperature_qv\" feature flag on the request. 
     * @deprecated
     */
    'temperatureUnit'?: Gridpoint12hForecastPeriodTemperatureUnitEnum;
    /**
     * If not null, indicates a non-diurnal temperature trend for the period (either rising temperature overnight, or falling temperature during the day) 
     */
    'temperatureTrend'?: Gridpoint12hForecastPeriodTemperatureTrendEnum | null;
    'probabilityOfPrecipitation'?: QuantitativeValue;
    'windSpeed'?: GridpointHourlyForecastPeriodWindSpeed;
    'windGust'?: GridpointHourlyForecastPeriodWindGust | null;
    /**
     * The prevailing direction of the wind for the period, using a 16-point compass.
     */
    'windDirection'?: Gridpoint12hForecastPeriodWindDirectionEnum;
    /**
     * A link to an icon representing the forecast summary.
     * @deprecated
     */
    'icon'?: string;
    /**
     * A brief textual forecast summary for the period.
     */
    'shortForecast'?: string;
    /**
     * A detailed textual forecast for the period.
     */
    'detailedForecast'?: string;
}

export const Gridpoint12hForecastPeriodTemperatureUnitEnum = {
    F: 'F',
    C: 'C'
} as const;

export type Gridpoint12hForecastPeriodTemperatureUnitEnum = typeof Gridpoint12hForecastPeriodTemperatureUnitEnum[keyof typeof Gridpoint12hForecastPeriodTemperatureUnitEnum];
export const Gridpoint12hForecastPeriodTemperatureTrendEnum = {
    Rising: 'rising',
    Falling: 'falling'
} as const;

export type Gridpoint12hForecastPeriodTemperatureTrendEnum = typeof Gridpoint12hForecastPeriodTemperatureTrendEnum[keyof typeof Gridpoint12hForecastPeriodTemperatureTrendEnum];
export const Gridpoint12hForecastPeriodWindDirectionEnum = {
    N: 'N',
    Nne: 'NNE',
    Ne: 'NE',
    Ene: 'ENE',
    E: 'E',
    Ese: 'ESE',
    Se: 'SE',
    Sse: 'SSE',
    S: 'S',
    Ssw: 'SSW',
    Sw: 'SW',
    Wsw: 'WSW',
    W: 'W',
    Wnw: 'WNW',
    Nw: 'NW',
    Nnw: 'NNW'
} as const;

export type Gridpoint12hForecastPeriodWindDirectionEnum = typeof Gridpoint12hForecastPeriodWindDirectionEnum[keyof typeof Gridpoint12hForecastPeriodWindDirectionEnum];