// src/openapi/generated/models/gridpoint12h-forecast.ts
/* tslint:disable */
/* eslint-disable */
/**
 * weather.gov API
 * weather.gov API
 *
 * The version of the OpenAPI document: 3.3.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Gridpoint12hForecastPeriod } from './gridpoint12h-forecast-period';
// May contain unused imports in some cases
// @ts-ignore
import type { GridpointForecastUnits } from './gridpoint-forecast-units';
// May contain unused imports in some cases
// @ts-ignore
import type { ISO8601Interval } from './iso8601-interval';
// May contain unused imports in some cases
// @ts-ignore
import type { JsonLdContext } from './json-ld-context';
// May contain unused imports in some cases
// @ts-ignore
import type { QuantitativeValue } from './quantitative-value';

/**
 * A multi-day forecast for a 2.5km grid square.
 */
export interface Gridpoint12hForecast {
    '@context'?: JsonLdContext;
    /**
     * A geometry represented in Well-Known Text (WKT) format.
     */
    'geometry'?: string | null;
    'units'?: GridpointForecastUnits;
    /**
     * The internal generator class used to create the forecast text (used for NWS debugging).
     */
    'forecastGenerator'?: string;
    /**
     * The time this forecast data was generated.
     */
    'generatedAt'?: string;
    /**
     * The last update time of the data this forecast was generated from.
     */
    'updateTime'?: string;
    'validTimes'?: ISO8601Interval;
    'elevation'?: QuantitativeValue;
    /**
     * An array of forecast periods.
     */
    'periods'?: Array<Gridpoint12hForecastPeriod>;
}