// src/openapi/generated/models/point.ts
/* tslint:disable */
/* eslint-disable */
/**
 * weather.gov API
 * weather.gov API
 *
 * The version of the OpenAPI document: 3.3.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { JsonLdContext } from './json-ld-context';
// May contain unused imports in some cases
// @ts-ignore
import type { NWSForecastOfficeId } from './nwsforecast-office-id';
// May contain unused imports in some cases
// @ts-ignore
import type { PointRelativeLocation } from './point-relative-location';

export interface Point {
    '@context'?: JsonLdContext;
    /**
     * A geometry represented in Well-Known Text (WKT) format.
     */
    'geometry'?: string | null;
    '@id'?: string;
    '@type'?: PointTypeEnum;
    'cwa'?: NWSForecastOfficeId;
    'forecastOffice'?: string;
    'gridId'?: NWSForecastOfficeId;
    'gridX'?: number;
    'gridY'?: number;
    'forecast'?: string;
    'forecastHourly'?: string;
    'forecastGridData'?: string;
    'observationStations'?: string;
    'relativeLocation'?: PointRelativeLocation;
    'forecastZone'?: string;
    'county'?: string;
    'fireWeatherZone'?: string;
    'timeZone'?: string;
    'radarStation'?: string;
}

export const PointTypeEnum = {
    WxPoint: 'wx:Point'
} as const;

export type PointTypeEnum = typeof PointTypeEnum[keyof typeof PointTypeEnum];