// src/openapi/generated/models/nwsforecast-office-id.ts
/* tslint:disable */
/* eslint-disable */
/**
 * weather.gov API
 * weather.gov API
 *
 * The version of the OpenAPI document: 3.3.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Three-letter identifier for a NWS office.
 */

export const NWSForecastOfficeId = {
    Akq: 'AKQ',
    Aly: 'ALY',
    Bgm: 'BGM',
    Box: 'BOX',
    Btv: 'BTV',
    Buf: 'BUF',
    Cae: 'CAE',
    Car: 'CAR',
    Chs: 'CHS',
    Cle: 'CLE',
    Ctp: 'CTP',
    Gsp: 'GSP',
    Gyx: 'GYX',
    Ilm: 'ILM',
    Iln: 'ILN',
    Lwx: 'LWX',
    Mhx: 'MHX',
    Okx: 'OKX',
    Pbz: 'PBZ',
    Phi: 'PHI',
    Rah: 'RAH',
    Rlx: 'RLX',
    Rnk: 'RNK',
    Abq: 'ABQ',
    Ama: 'AMA',
    Bmx: 'BMX',
    Bro: 'BRO',
    Crp: 'CRP',
    Epz: 'EPZ',
    Ewx: 'EWX',
    Ffc: 'FFC',
    Fwd: 'FWD',
    Hgx: 'HGX',
    Hun: 'HUN',
    Jan: 'JAN',
    Jax: 'JAX',
    Key: 'KEY',
    Lch: 'LCH',
    Lix: 'LIX',
    Lub: 'LUB',
    Lzk: 'LZK',
    Maf: 'MAF',
    Meg: 'MEG',
    Mfl: 'MFL',
    Mlb: 'MLB',
    Mob: 'MOB',
    Mrx: 'MRX',
    Ohx: 'OHX',
    Oun: 'OUN',
    Shv: 'SHV',
    Sjt: 'SJT',
    Sju: 'SJU',
    Tae: 'TAE',
    Tbw: 'TBW',
    Tsa: 'TSA',
    Abr: 'ABR',
    Apx: 'APX',
    Arx: 'ARX',
    Bis: 'BIS',
    Bou: 'BOU',
    Cys: 'CYS',
    Ddc: 'DDC',
    Dlh: 'DLH',
    Dmx: 'DMX',
    Dtx: 'DTX',
    Dvn: 'DVN',
    Eax: 'EAX',
    Fgf: 'FGF',
    Fsd: 'FSD',
    Gid: 'GID',
    Gjt: 'GJT',
    Gld: 'GLD',
    Grb: 'GRB',
    Grr: 'GRR',
    Ict: 'ICT',
    Ilx: 'ILX',
    Ind: 'IND',
    Iwx: 'IWX',
    Jkl: 'JKL',
    Lbf: 'LBF',
    Lmk: 'LMK',
    Lot: 'LOT',
    Lsx: 'LSX',
    Mkx: 'MKX',
    Mpx: 'MPX',
    Mqt: 'MQT',
    Oax: 'OAX',
    Pah: 'PAH',
    Pub: 'PUB',
    Riw: 'RIW',
    Sgf: 'SGF',
    Top: 'TOP',
    Unr: 'UNR',
    Boi: 'BOI',
    Byz: 'BYZ',
    Eka: 'EKA',
    Fgz: 'FGZ',
    Ggw: 'GGW',
    Hnx: 'HNX',
    Lkn: 'LKN',
    Lox: 'LOX',
    Mfr: 'MFR',
    Mso: 'MSO',
    Mtr: 'MTR',
    Otx: 'OTX',
    Pdt: 'PDT',
    Pih: 'PIH',
    Pqr: 'PQR',
    Psr: 'PSR',
    Rev: 'REV',
    Sew: 'SEW',
    Sgx: 'SGX',
    Slc: 'SLC',
    Sto: 'STO',
    Tfx: 'TFX',
    Twc: 'TWC',
    Vef: 'VEF',
    Aer: 'AER',
    Afc: 'AFC',
    Afg: 'AFG',
    Ajk: 'AJK',
    Alu: 'ALU',
    Gum: 'GUM',
    Hpa: 'HPA',
    Hfo: 'HFO',
    Ppg: 'PPG',
    Pqe: 'PQE',
    Pqw: 'PQW',
    Stu: 'STU',
    Nh1: 'NH1',
    Nh2: 'NH2',
    Ona: 'ONA',
    Onp: 'ONP'
} as const;

export type NWSForecastOfficeId = typeof NWSForecastOfficeId[keyof typeof NWSForecastOfficeId];